[rtla] Implement housekeeping CPU sliding scale logic Add a function populate_numa_info that generated information per NUMA node. Generates cores per NUMA node, which CPUs per NUMA node are housekeeping, and which CPU per NUMA node are workers. This works across ranges of cpusets.
Add variable HOUSEKEEPING_PER_NUMA which is used to determine the
sliding scale. It defaults to 0. When no housekeeping is specified,
use the provided cpuset as measurement threads. HOUSEKEEPING_PER_NUMA
overrides all other settings and need to match the CPUSET provided by
the runtime.

Lastly, add ALLCPUSET to be used to indicate to the container to use
all available CPUs. This is due to a bug in RHEL9 where when no cpuset
is provided to the runtime, /proc/self/status reports the cpuset being
CPU0 instead of all CPUs. This is for when no cpuset is provided by the
runtime which previously would get bound to CPU0 instead of all
of the actual CPUs available to the container. If a CPUSET is passed
by the runtime, then this variable is ignored. This can be deprecated
once the bug is fixed as it is in Fedora

Signed-off-by: Chris White <chwhite@redhat.com>